Update check_ip
Switched variable names around to make more sense
Added a check for the new IP address. If the string was empty, it used to email the current IP address. Now, if it fails to get the new IP address it will simply try again later based on the crontab config.